refactor(insights): scope traffic sources to builder-insights only
