+++ app/(home)/hackathons/[id]/admin-panel/page.tsx
@@ -1,5 +1,4 @@
 import HackathonForm from "@/components/hackathons/admin-panel/HackathonForm";
-import TopTrafficSourcesCard from "@/components/hackathons/admin-panel/TopTrafficSourcesCard";
 import { getHackathon } from "@/server/services/hackathons";
 import { redirect } from "next/navigation";
 import { getAuthSession } from "@/lib/auth/authSession";
@@ -21,8 +20,7 @@ export default async function HackathonAdminPanel({
   if (!hackathon) redirect("/events");
 
   return (
-    <main className="container  relative px-2 py-4 lg:py-16 flex flex-col gap-6">
-      <TopTrafficSourcesCard hackathonId={id} />
+    <main className="container  relative px-2 py-4 lg:py-16">
       <div className="border border-zinc-800 shadow-sm bg-zinc-950 rounded-md">
         <HackathonForm initialData={hackathon} isEditing={true} />
       </div>
