+++ server/services/hackathonTrafficSources.ts
@@ -13,12 +13,17 @@ const HOGQL_HOST_FILTER =
  * only be attributed at the channel level. UTM-tagged links keep full
  * granularity via the first branch (use utm_content for the poster handle).
  *
- * Priority: explicit UTM → broad channel → bare domain → "Direct".
+ * Priority: explicit UTM (excluding PostHog's '$direct' sentinel) →
+ * sign-in/OAuth redirects → broad channel → bare domain → "Direct".
  */
 const SOURCE_BUCKET_EXPR = `
   multiIf(
-    notEmpty(properties.utm_source),
+    notEmpty(properties.utm_source) AND properties.utm_source != '$direct',
       concat(properties.utm_source, ' / ', coalesce(properties.utm_campaign, '(no campaign)')),
+    properties.$referring_domain IN (
+      'accounts.google.com', 'login.microsoftonline.com', 'github.com'
+    ),
+      'Sign-in redirect',
     properties.$referring_domain IN ('x.com', 'twitter.com', 't.co'),
       'X (untagged)',
     properties.$referring_domain = 'linkedin.com'
@@ -69,56 +74,6 @@ export interface TopTrafficSourcesOptions {
   days?: number;
   /** Number of source buckets to return. Default 3. */
   limit?: number;
-}
-
-/**
- * Top traffic sources for a single hackathon, derived from PostHog `$pageview`
- * events. Returns at most `limit` rows ordered by unique visitors desc.
- *
- * Matches events either by the new sticky `hackathon_id` property OR by URL
- * pattern (`/hackathons/<id>...`), so attribution still works for traffic that
- * arrived before the property-enrichment shipped.
- *
- * Never throws — returns [] on missing config or any error, matching the
- * `runHogQL` contract.
- */
-export async function getTopHackathonTrafficSources(
-  hackathonId: string,
-  { days = 90, limit = 3 }: TopTrafficSourcesOptions = {},
-): Promise<HackathonTrafficSource[]> {
-  if (!isSafeHackathonId(hackathonId)) return [];
-
-  const safeDays = Math.max(1, Math.min(365, Math.floor(days)));
-  const safeLimit = Math.max(1, Math.min(20, Math.floor(limit)));
-
-  const query = `
-    SELECT
-      ${SOURCE_BUCKET_EXPR} AS source,
-      count(DISTINCT distinct_id) AS visitors,
-      countIf(properties.$pathname LIKE '%/registration-form%') AS reachedRegister
-    FROM events
-    WHERE event = '$pageview'
-      AND ${HOGQL_HOST_FILTER}
-      AND (
-        properties.hackathon_id = '${hackathonId}'
-        OR properties.$pathname LIKE '/hackathons/${hackathonId}%'
-      )
-      AND timestamp >= now() - INTERVAL ${safeDays} DAY
-    GROUP BY source
-    ORDER BY visitors DESC
-    LIMIT ${safeLimit}
-  `.trim();
-
-  const rows = await runHogQL<RawRow>({
-    projectId: POSTHOG_BUILDER_HUB_PROJECT_ID,
-    query,
-  });
-
-  return rows.map((row) => ({
-    source: row.source ?? "Direct",
-    visitors: toNumber(row.visitors),
-    reachedRegister: toNumber(row.reachedRegister),
-  }));
 }
 
 interface BatchRow extends RawRow {
